* Added option: CLICON_CLI_BUF_START and CLICON_CLI_BUF_THRESHOLD so you can change the start and     threshold of quadratic and linear growth of CLIgen buffers (cbuf:s) * Memory footprint   * Do not autopopulate namespace cache, instead use on-demand, see `xml2ns()`.   * Set CBUF start level to 256 (`CLICON_CLI_BUF_START` option)   * Reduced xml child vector default size from 4 to 1 with quadratic growoth to 64K then linear

+++ test/test_nacm_module_read.sh
@@ -139,14 +139,16 @@ fi
 new "waiting"
 wait_backend
 
-new "kill old restconf daemon"
-sudo pkill -u $wwwuser -f clixon_restconf
+if [ $RC -ne 0 ]; then
+    new "kill old restconf daemon"
+    sudo pkill -u $wwwuser -f clixon_restconf
 
-new "start restconf daemon (-a is enable basic authentication)"
-start_restconf -f $cfg -- -a
+    new "start restconf daemon (-a is enable basic authentication)"
+    start_restconf -f $cfg -- -a
 
-new "waiting"
-wait_restconf
+    new "waiting"
+    wait_restconf
+fi
 
 new "auth set authentication config"
 expecteof "$clixon_netconf -qf $cfg" 0 "<rpc><edit-config><target><candidate/></target><config>$RULES</config></edit-config></rpc>]]>]]>" "^<rpc-reply><ok/></rpc-reply>]]>]]>$"
@@ -271,9 +273,10 @@ new "guest read state fail"
 expecteq "$(curl -u guest:bar -sS -X GET http://localhost/restconf/data/clixon-example:state)" 0 '{"ietf-restconf:errors":{"error":{"error-type":"application","error-tag":"access-denied","error-severity":"error","error-message":"default deny"}}}'
 
 
-new "Kill restconf daemon"
-stop_restconf 
-
+if [ $RC -ne 0 ]; then
+    new "Kill restconf daemon"
+    stop_restconf 
+fi
 
 if [ $BE -eq 0 ]; then
     exit # BE
